Comienzo de sincronizacion entre proceso ClientsGenerator y Receptionists

+++ src/data/Config.cpp
@@ -90,7 +90,7 @@ vector<Food> Config::getFoodsFromLine(string line) {
         string foodCost;
         getline(f, foodName, ',');
         getline(f, foodCost, ',');
-        int cost = std::stoi(foodCost);
+        float cost = std::stof(foodCost);
         foodsVector.push_back(Food(foodName, cost));
     }
     return foodsVector;
